backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
firebase-admin==6.4.0
openai==1.12.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.3.0
